Move stack/storage to where they're used
